+++ lib/src/hosts/azure_host.dart
@@ -51,6 +51,9 @@ class AzureHost extends HostAdapter {
   String get configKey => 'azure';
 
   @override
+  bool get supportsAllInOneWeb => true;
+
+  @override
   List<DatabaseProvider> get supportedDatabases => const [
         DatabaseProvider.none,
         DatabaseProvider.neon,